feat: dev mode for assets; fonts in SVGs
