+++ src/module/Ui/templates/authorization/role/show.phtml
@@ -48,7 +48,7 @@
                         ]
                     ]); ?>
             <?php endif; ?>
-            <?php echo $user->getUsername() ?>
+            <?php echo $user->getUsername(); ?>
         </li>
     <?php endforeach; ?>
 </ul>
