+++ src/module/Ui/templates/editor/layout.phtml
@@ -28,7 +28,7 @@ echo $this->doctype();
 <html class="fuelux" lang="<?php echo $this->currentLanguage(); ?>"> <!--<![endif]-->
 <head>
     <?php print $this->partial('layout/partials/head');?>
-    <link href="<?php echo $this->assetsHost(); ?>legacy-editor.css" rel="stylesheet">
+    <link href="<?php echo $this->assets()->getBundle('legacy-editor.css'); ?>" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Open+Sans:400italic,400,700,600|Droid+Serif" rel="stylesheet">
 </head>
 <body>
