+++ src/module/Ui/templates/editor/partials/scripts.phtml
@@ -24,5 +24,5 @@
 <?php echo $this->partial('layout/partials/mathjax'); ?>
 <?php echo $this->headScript(); ?>
 <?php echo $this->inlineScript(); ?>
-<script src="<?php echo $this->assetsHost(); ?>legacy-editor.js" type="text/javascript"></script>
+<script src="<?php echo $this->assets()->getBundle('legacy-editor.js'); ?>" type="text/javascript"></script>
 <?php echo $this->tracking(); ?>
