+++ src/module/Ui/templates/error/404.phtml
@@ -73,8 +73,8 @@ $this->headTitle('That page does not exist!');
     <?php if (isset($this->controller) && $this->controller): ?>
 
         <dl>
-            <dt><?php echo $this->translate('Controller') ?>:</dt>
-            <dd><?php echo $this->escapeHtml($this->controller) ?>
+            <dt><?php echo $this->translate('Controller'); ?>:</dt>
+            <dd><?php echo $this->escapeHtml($this->controller); ?>
                 <?php
                 if (isset($this->controller_class) && $this->controller_class && $this->controller_class != $this->controller
                 ) {
@@ -93,21 +93,21 @@ $this->headTitle('That page does not exist!');
 
         <?php if (isset($this->exception) && $this->exception instanceof Exception): ?>
             <hr/>
-            <h2><?php echo $this->translate('Additional information') ?>:</h2>
+            <h2><?php echo $this->translate('Additional information'); ?>:</h2>
             <h3><?php echo get_class($this->exception); ?></h3>
             <dl>
-                <dt><?php echo $this->translate('File') ?>:</dt>
+                <dt><?php echo $this->translate('File'); ?>:</dt>
                 <dd>
-                    <pre class="prettyprint linenums"><?php echo $this->exception->getFile() ?>
-                        :<?php echo $this->exception->getLine() ?></pre>
+                    <pre class="prettyprint linenums"><?php echo $this->exception->getFile(); ?>
+                        :<?php echo $this->exception->getLine(); ?></pre>
                 </dd>
-                <dt><?php echo $this->translate('Message') ?>:</dt>
+                <dt><?php echo $this->translate('Message'); ?>:</dt>
                 <dd>
-                    <pre class="prettyprint linenums"><?php echo $this->exception->getMessage() ?></pre>
+                    <pre class="prettyprint linenums"><?php echo $this->exception->getMessage(); ?></pre>
                 </dd>
-                <dt><?php echo $this->translate('Stack trace') ?>:</dt>
+                <dt><?php echo $this->translate('Stack trace'); ?>:</dt>
                 <dd>
-                    <pre class="prettyprint linenums"><?php echo $this->exception->getTraceAsString() ?></pre>
+                    <pre class="prettyprint linenums"><?php echo $this->exception->getTraceAsString(); ?></pre>
                 </dd>
             </dl>
             <?php
@@ -115,24 +115,24 @@ $this->headTitle('That page does not exist!');
             if ($e) :
                 ?>
                 <hr/>
-                <h2><?php echo $this->translate('Previous exceptions') ?>:</h2>
+                <h2><?php echo $this->translate('Previous exceptions'); ?>:</h2>
                 <ul class="unstyled">
                     <?php while ($e) : ?>
                         <li>
                             <h3><?php echo get_class($e); ?></h3>
                             <dl>
-                                <dt><?php echo $this->translate('File') ?>:</dt>
+                                <dt><?php echo $this->translate('File'); ?>:</dt>
                                 <dd>
-                                    <pre class="prettyprint linenums"><?php echo $e->getFile() ?>
-                                        :<?php echo $e->getLine() ?></pre>
+                                    <pre class="prettyprint linenums"><?php echo $e->getFile(); ?>
+                                        :<?php echo $e->getLine(); ?></pre>
                                 </dd>
-                                <dt><?php echo $this->translate('Message') ?>:</dt>
+                                <dt><?php echo $this->translate('Message'); ?>:</dt>
                                 <dd>
-                                    <pre class="prettyprint linenums"><?php echo $e->getMessage() ?></pre>
+                                    <pre class="prettyprint linenums"><?php echo $e->getMessage(); ?></pre>
                                 </dd>
-                                <dt><?php echo $this->translate('Stack trace') ?>:</dt>
+                                <dt><?php echo $this->translate('Stack trace'); ?>:</dt>
                                 <dd>
-                                    <pre class="prettyprint linenums"><?php echo $e->getTraceAsString() ?></pre>
+                                    <pre class="prettyprint linenums"><?php echo $e->getTraceAsString(); ?></pre>
                                 </dd>
                             </dl>
                         </li>
