+++ src/module/Ui/templates/error/index.phtml
@@ -33,20 +33,20 @@
 
 <?php if(isset($this->exception) && $this->exception instanceof Exception): ?>
 <div class="well">
-<h2><?php echo $this->translate('Additional information') ?>:</h2>
+<h2><?php echo $this->translate('Additional information'); ?>:</h2>
 <h3><?php echo get_class($this->exception); ?></h3>
 <dl>
-    <dt><?php echo $this->translate('File') ?>:</dt>
+    <dt><?php echo $this->translate('File'); ?>:</dt>
     <dd>
-        <pre class="prettyprint linenums"><?php echo $this->exception->getFile() ?>:<?php echo $this->exception->getLine() ?></pre>
+        <pre class="prettyprint linenums"><?php echo $this->exception->getFile(); ?>:<?php echo $this->exception->getLine(); ?></pre>
     </dd>
-    <dt><?php echo $this->translate('Message') ?>:</dt>
+    <dt><?php echo $this->translate('Message'); ?>:</dt>
     <dd>
-        <pre class="prettyprint linenums"><?php echo $this->exception->getMessage() ?></pre>
+        <pre class="prettyprint linenums"><?php echo $this->exception->getMessage(); ?></pre>
     </dd>
-    <dt><?php echo $this->translate('Stack trace') ?>:</dt>
+    <dt><?php echo $this->translate('Stack trace'); ?>:</dt>
     <dd>
-        <pre class="prettyprint linenums"><?php echo $this->exception->getTraceAsString() ?></pre>
+        <pre class="prettyprint linenums"><?php echo $this->exception->getTraceAsString(); ?></pre>
     </dd>
 </dl>
 <?php
@@ -54,23 +54,23 @@
     if ($e) :
 ?>
 <hr/>
-<h2><?php echo $this->translate('Previous exceptions') ?>:</h2>
+<h2><?php echo $this->translate('Previous exceptions'); ?>:</h2>
 <ul class="unstyled">
     <?php while($e) : ?>
     <li>
         <h3><?php echo get_class($e); ?></h3>
         <dl>
-            <dt><?php echo $this->translate('File') ?>:</dt>
+            <dt><?php echo $this->translate('File'); ?>:</dt>
             <dd>
-                <pre class="prettyprint linenums"><?php echo $e->getFile() ?>:<?php echo $e->getLine() ?></pre>
+                <pre class="prettyprint linenums"><?php echo $e->getFile(); ?>:<?php echo $e->getLine(); ?></pre>
             </dd>
-            <dt><?php echo $this->translate('Message') ?>:</dt>
+            <dt><?php echo $this->translate('Message'); ?>:</dt>
             <dd>
-                <pre class="prettyprint linenums"><?php echo $e->getMessage() ?></pre>
+                <pre class="prettyprint linenums"><?php echo $e->getMessage(); ?></pre>
             </dd>
-            <dt><?php echo $this->translate('Stack trace') ?>:</dt>
+            <dt><?php echo $this->translate('Stack trace'); ?>:</dt>
             <dd>
-                <pre class="prettyprint linenums"><?php echo $e->getTraceAsString() ?></pre>
+                <pre class="prettyprint linenums"><?php echo $e->getTraceAsString(); ?></pre>
             </dd>
         </dl>
     </li>
@@ -83,7 +83,7 @@
 
 <?php else: ?>
 
-<h3><?php echo $this->translate('No Exception available') ?></h3>
+<h3><?php echo $this->translate('No Exception available'); ?></h3>
 
 </div>
