+++ src/module/Ui/templates/layout/2-col.phtml
@@ -71,7 +71,7 @@ $pageClass = join(' ', $pageClasses);
 </div>
 <?php echo $this->partial('layout/partials/footer'); ?>
 <?php echo $this->partial('layout/partials/default-modal'); ?>
-<?php echo $this->placeholder('modals') ?>
+<?php echo $this->placeholder('modals'); ?>
 <?php echo $this->tracking(); ?>
 <?php echo $this->partial('layout/partials/noscript'); ?>
 </body>
