+++ src/module/Ui/templates/layout/en/serlo-home.phtml
@@ -36,7 +36,7 @@ echo $this->doctype();
                 <div id="mobile-nav-toggle">
 
                     <button>
-                        <span class="sr-only"><?php echo $this->translate('Toggle navigation') ?></span>
+                        <span class="sr-only"><?php echo $this->translate('Toggle navigation'); ?></span>
                         <i class="fa fa-fw fa-bars"></i>
                     </button>
 
@@ -128,7 +128,7 @@ echo $this->doctype();
                 </div>
 
                 <div class="home-row home-principle-row home-column col-md-12 col-lg-6 col-lg-offset-3">
-                    <img src="https://assets.serlo.org/athene2/en/home/about-serlo.svg" alt="About Serlo">
+                    <object data="<?php echo $this->assets()->getAsset('en/home/about-serlo.svg'); ?>"></object>
                 </div>
 
                 <div class="home-calls-wrapper clearfix">
@@ -137,7 +137,7 @@ echo $this->doctype();
                             <h2>Serlo is the Wikipedia for Learning<h2>
                         </div>
                         <div class="extra-icon col-xs-12 col-lg-2">
-                            <img src="https://assets.serlo.org/athene2/home/participate.svg" alt="together">
+                            <img src="<?php echo $this->assets()->getAsset('home/participate.svg'); ?>" alt="together">
                         </div>
                         <div class="col-sm-6 col-sm-offset-3 col-md-4 col-md-offset-1 col-lg-offset-0 col-lg-3 home-calls-column calls-column-1">
                             <p>We are looking for teachers and enthusiastic educators who are passionate about their subject.
