+++ src/module/Ui/templates/layout/partials/footer.phtml
@@ -25,7 +25,7 @@
     <footer id="footer" class="home-row">
         <div id="mission-statement" class="col-lg-3 pull-right-lg">
             <div id="mission-statement-claim">
-                <h2 id="mission-statement-head"><?php echo $this->translate('Serlo.org is the Wikipedia for learning.') ?></h2>
+                <h2 id="mission-statement-head"><?php echo $this->translate('Serlo.org is the Wikipedia for learning.'); ?></h2>
                 <p><?php echo $this->translate('We are a community of visionaries working tirelessly to make great education freely available to everyone.'); ?></p>
                 <a href="/serlo"><?php echo $this->translate('Learn more'); ?> <i class="fa fa-arrow-circle-right"></i>
                 </a>
@@ -33,14 +33,14 @@
             <div id="mission-statement-actions" class="row">
                 <div class="mission-statement-action col-xs-6">
                     <a href="/mitmachen">
-                        <img src="https://assets.serlo.org/athene2/footer/participate.svg"/>
-                        <p><?php echo $this->translate('Join the cause') ?></p>
+                        <img src="<?php echo $this->assets()->getAsset('footer/participate.svg'); ?>"/>
+                        <p><?php echo $this->translate('Join the cause'); ?></p>
                     </a>
                 </div>
                 <div class="mission-statement-action col-xs-6">
                     <a href="/spenden">
                         <?php echo $this->partial('layout/de/home_donate');?>
-                        <p><?php echo $this->translate('Donate') ?></p>
+                        <p><?php echo $this->translate('Donate'); ?></p>
                     </a>
                 </div>
             </div>
@@ -56,7 +56,7 @@
             <nav>
                 <div class="row">
                     <div class="col-lg-2 col-sm-3">
-                        <div class="footer-title"><?php echo $this->translate('General') ?></div>
+                        <div class="footer-title"><?php echo $this->translate('General'); ?></div>
                         <?php echo $this->navigation('footer_left_navigation')
                             ->menu()
                             ->setPartial('layout/navigation/vertical')
@@ -64,7 +64,7 @@
                         ?>
                     </div>
                     <div class="col-lg-2 col-sm-3">
-                        <div class="footer-title"><?php echo $this->translate('Subjects') ?></div>
+                        <div class="footer-title"><?php echo $this->translate('Subjects'); ?></div>
                         <?php echo $this->navigation('subject_navigation')
                             ->menu()
                             ->setPartial('layout/navigation/vertical')
@@ -80,7 +80,7 @@
                         ?>
                     </div>
                     <div class="col-lg-4 col-sm-3">
-                        <div class="footer-title"><?php echo $this->translate('Legal terms') ?></div>
+                        <div class="footer-title"><?php echo $this->translate('Legal terms'); ?></div>
                         <?php echo $this->navigation('footer_right_navigation')
                             ->menu()
                             ->setPartial('layout/navigation/vertical')
