+++ src/module/Ui/templates/layout/partials/scripts.phtml
@@ -24,5 +24,5 @@
 <?php echo $this->partial('layout/partials/mathjax'); ?>
 <?php echo $this->headScript(); ?>
 <?php echo $this->inlineScript(); ?>
-<script src="<?php echo $this->assetsHost(); ?>main.js" type="text/javascript"></script>
+<script src="<?php echo $this->assets()->getBundle('main.js'); ?>" type="text/javascript"></script>
 <?php echo $this->partial('layout/partials/popup'); ?>
