+++ src/module/Ui/templates/layout/partials/styles.phtml
@@ -22,4 +22,4 @@
  */
 ?>
 <?php echo $this->headStyle(); ?>
-<link href="<?php echo $this->assetsHost(); ?>main.css" rel="stylesheet">
+<link href="<?php echo $this->assets()->getBundle('main.css'); ?>" rel="stylesheet">
